+++ tests/client/__init__.py
@@ -1,5 +1,7 @@
 from tests.client.client import (
     Client,
+    FlakyTest,
+    HealthCheckFailure,
     assume,
     collection,
     generate_from_schema,
@@ -12,6 +14,8 @@ from tests.client.protocol import ClientConnection
 __all__ = [
     "Client",
     "ClientConnection",
+    "FlakyTest",
+    "HealthCheckFailure",
     "assume",
     "collection",
     "generate_from_schema",